# frontend/openapi-snapshot.json
{
  "components": {
    "schemas": {
      "AISearchQuery": {
        "description": "Input: free text + optional conversation history.",
        "properties": {
          "conversation_history": {
            "description": "Previous turns for context (max 6)",
            "items": {
              "$ref": "#/components/schemas/ConversationTurn"
            },
            "maxItems": 6,
            "title": "Conversation History",
            "type": "array"
          },
          "query": {
            "maxLength": 400,
            "minLength": 1,
            "title": "Query",
            "type": "string"
          }
        },
        "required": [
          "query"
        ],
        "title": "AISearchQuery",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PaginatedPassengerRequestsResponse": {
        "description": "Paginated passenger requests for GET /passengers/me.",
        "properties": {
          "has_more": {
            "default": false,
            "title": "Has More",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/PassengerRequestResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "title": "PaginatedPassengerRequestsResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "LinkUp API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/passenger/passengers/": {
      "post": {
        "operationId": "create_new_request_api_v1_passenger_passengers__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PassengerRequestCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PassengerRequestWithMatches"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "רישום נוסע לטרמפ (יצירת בקשה קבועה)",
        "tags": [
          "Passenger",
          "Passenger"
        ]
      }
    },
    "/api/v1/passenger/passengers/ai-parse-search": {
      "post": {
        "description": "Parse Hebrew/English free text into structured pickup/destination/time/radius.\nDoes not run search — client fills the form and calls GET /search-rides.",
        "operationId": "ai_parse_search_api_v1_passenger_passengers_ai_parse_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AISearchQuery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AISearchResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "ניתוח חיפוש נסיעה בטקסט חופשי (AI)",
        "tags": [
          "Passenger",
          "Passenger"
        ]
      }
    },
    "/api/v1/passenger/passengers/all": {
      "get": {
        "description": "Return all rides in the system; optional status filter.",
        "operationId": "get_all_rides_admin_api_v1_passenger_passengers_all_get",
        "parameters": [
          {
            "description": "סנן לפי סטטוס: open, cancelled, completed",
            "in": "query",
            "name": "filter_status",
            "required": false,
            "schema": {
              "description": "סנן לפי סטטוס: open, cancelled, completed",
              "title": "Filter Status",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RideResponse"
                  },
                  "title": "Response Get All Rides Admin Api V1 Passenger Passengers All Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "תצוגת כל הנסיעות (ניהול ובקרה)",
        "tags": [
          "Passenger",
          "Passenger"
        ]
      }
    },
    "/api/v1/passenger/passengers/me": {
      "get": {
        "description": "List the current user's passenger requests (paginated).",
        "operationId": "get_my_requests_api_v1_passenger_passengers_me_get",
        "parameters": [
          {
            "description": "סנן לפי סטטוס: pending, approved, cancelled, matched, expired, completed, rejected",
            "in": "query",
            "name": "request_status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "סנן לפי סטטוס: pending, approved, cancelled, matched, expired, completed, rejected",
              "title": "Request Status"
            }
          },
          {
            "description": "cursor אטום להמשך העמוד הבא",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "cursor אטום להמשך העמוד הבא",
              "title": "Cursor"
            }
          },
          {
            "description": "פריטים לעמוד",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "פריטים לעמוד",
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedPassengerRequestsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get My Requests",
        "tags": [
          "Passenger",
          "Passenger"
        ]
      }
    },
    "/api/v1/passenger/passengers/request-ride-from-search": {
      "post": {
        "description": "Join request from search results.\nOptional Idempotency-Key header prevents duplicate bookings on retry/double-click.\nSame key + different body → 422. Same key + same body → returns original result.",
        "operationId": "request_ride_from_search_api_v1_passenger_passengers_request_ride_from_search_post",
        "parameters": [
          {
            "description": "UUID ייחודי לכל כוונת הצטרפות — מניע כפילות בלחיצה כפולה",
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "UUID ייחודי לכל כוונת הצטרפות — מניע כפילות בלחיצה כפולה",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestRideFromSearch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "שלח בקשה להצטרפות לנסיעה מתוך תוצאות חיפוש",
        "tags": [
          "Passenger",
          "Passenger"
        ]
      }
    },
    "/api/v1/passenger/passengers/search-rides": {
      "get": {
        "operationId": "search_available_rides_api_v1_passenger_passengers_search_rides_get",
        "parameters": [
          {
            "in": "query",
            "name": "pickup_name",
            "required": true,
            "schema": {
              "title": "Pickup Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "destination_name",
            "required": true,
            "schema": {
              "title": "Destination Name",
              "type": "string"
            }
          },
          {
            "description": "רדיוס חיפוש בקילומטרים (אחיד עם יצירת בקשה)",
            "in": "query",
            "name": "search_radius",
            "required": false,
            "schema": {
              "default": 1.0,
              "description": "רדיוס חיפוש בקילומטרים (אחיד עם יצירת בקשה)",
              "maximum": 50,
              "minimum": 0.1,
              "title": "Search Radius",
              "type": "number"
            }
          },
          {
            "description": "יום מלא Asia/Jerusalem — הדדי ל־departure_time / departure_time_to",
            "in": "query",
            "name": "departure_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "יום מלא Asia/Jerusalem — הדדי ל־departure_time / departure_time_to",
              "title": "Departure Date"
            }
          },
          {
            "description": "±2 שעות; או תחילת טווח עם departure_time_to",
            "in": "query",
            "name": "departure_time",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "±2 שעות; או תחילת טווח עם departure_time_to",
              "title": "Departure Time"
            }
          },
          {
            "description": "סוף טווח כולל (דורש departure_time)",
            "in": "query",
            "name": "departure_time_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "סוף טווח כולל (דורש departure_time)",
              "title": "Departure Time To"
            }
          },
          {
            "description": "רדיוס יעד בקילומטרים (אופציונלי)",
            "in": "query",
            "name": "destination_radius",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 50,
                  "minimum": 0.1,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "רדיוס יעד בקילומטרים (אופציונלי)",
              "title": "Destination Radius"
            }
          },
          {
            "description": "כמות תוצאות",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "description": "כמות תוצאות",
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "opaque cursor",
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "opaque cursor",
              "title": "After"
            }
          },
          {
            "in": "query",
            "name": "group_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Group Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RideSearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "חיפוש טרמפים (ללא שמירת בקשה; שמירת התראה ב-POST /)",
        "tags": [
          "Passenger",
          "Passenger"
        ]
      }
    },
    "/api/v1/passenger/passengers/{request_id}/cancel": {
      "delete": {
        "description": "Cancel the request and release all seat holds against drivers (request owner only).",
        "operationId": "cancel_request_api_v1_passenger_passengers__request_id__cancel_delete",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "ביטול בקשת נסיעה ושחרור שריונים",
        "tags": [
          "Passenger",
          "Passenger"
        ]
      }
    },
    "/api/v1/passenger/passengers/{request_id}/matches": {
      "get": {
        "operationId": "get_latest_matches_api_v1_passenger_passengers__request_id__matches_get",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RideResponse"
                  },
                  "title": "Response Get Latest Matches Api V1 Passenger Passengers  Request Id  Matches Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "שליפת התאמות עדכניות לבקשה קיימת",
        "tags": [
          "Passenger",
          "Passenger"
        ]
      }
    },
...
    "/api/v1/rides/": {
      "post": {
        "operationId": "create_new_ride_api_v1_rides__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RideCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RideResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create New Ride",
        "tags": [
          "Rides"
        ]
      }
    },
...
    "/api/v1/rides/{ride_id}/cancel": {
      "delete": {
        "operationId": "cancel_ride_api_v1_rides__ride_id__cancel_delete",
        "parameters": [
          {
            "in": "path",
            "name": "ride_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Ride Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Cancel Ride",
        "tags": [
          "Rides"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Local",
      "url": "http://127.0.0.1:8000"
    }
  ]
}